Finish responsive menu
